scripts/word_report/requirements.txt
cycler==0.10.0
docxtpl==0.11.3
Jinja2==2.11.2
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
six==1.15.0
